Data.Bounds

Description

Provides various typeclasses for dealing with bounded and unbounded types.

Bounded

class LowerBounded a where

Names the lower limit of a type. Types that also have a Bounded instance should define lowerBound === minBound. This can be derived with the anyclass strategy.

Examples

Expand

>>> -- -XDeriveAnyClass, -XDerivingStrategies
>>> data Foo = Foo1 | Foo2 deriving stock (Bounded, Show) deriving anyclass LowerBounded
>>> lowerBound @Foo
Foo1

Since: 0.1

Minimal complete definition

Nothing

Methods

lowerBound :: a

Since: 0.1

default lowerBound :: Bounded a => a

class UpperBounded a where

Names the upper limit of a type. Types that also have a Bounded instance should define upperBound === maxBound. This can be derived with the anyclass strategy.

Examples

Expand

>>> -- -XDeriveAnyClass, -XDerivingStrategies
>>> data Foo = Foo1 | Foo2 deriving stock (Bounded, Show) deriving anyclass UpperBounded
>>> upperBound @Foo
Foo2

Since: 0.1

Minimal complete definition

Nothing

Methods

upperBound :: a

Since: 0.1

default upperBound :: Bounded a => a

newtype Bounds a

Newtype that gives UpperBounded and LowerBounded instances to types that only have a Bounded instance.

Examples

Expand

>>> newtype Foo = MkFoo Int8 deriving stock (Bounded, Show)
>>> upperBound @(Bounds Foo)
MkBounds {unBounds = MkFoo 127}

>>> lowerBound @(Bounds Foo)
MkBounds {unBounds = MkFoo (-128)}

Since: 0.1

Constructors

MkBounds

Fields

unBounds :: a

Since: 0.1

Boundless

class LowerBoundless a

Types that have no lower bound.

Since: 0.1

class UpperBoundless a

Types that have no upper bound.

Since: 0.1

class (LowerBoundless a, UpperBoundless a) => Boundless a

Types that are unbounded above and below.

Since: 0.1

AnyBounded

class AnyLowerBounded a where

Class for types that may have a lower bound. The intention is that bounded types (e.g. Int8) return Just minBound whereas unbounded types (e.g. Integer) return Nothing.

Since: 0.1

Minimal complete definition

Nothing

Methods

someLowerBound :: Maybe a

Retrieves the lower bound.

Since: 0.1

default someLowerBound :: LowerBounded a => Maybe a

class AnyUpperBounded a where

Class for types that may have an upper bound. The intention is that bounded types (e.g. Int8) return Just maxBound whereas unbounded types (e.g. Integer) return Nothing.

Since: 0.1

Minimal complete definition

Nothing

Methods

someUpperBound :: Maybe a

Retrieves the upper bound.

Since: 0.1

default someUpperBound :: UpperBounded a => Maybe a